producer/Weather/requirements.txt
requests
kafka-python
python-dotenv
sentry-sdk
psycopg2-binary